+++ modules/database/src/ioc/db/dbAccess.c
@@ -756,7 +756,7 @@ long dbValueSize(short dbr_type)
         sizeof(epicsFloat64),        /* DOUBLE       */
         sizeof(epicsEnum16)};        /* ENUM         */
 
-    if(dbr_type>=NELEMENTS(size))
+    if((size_t)dbr_type>=NELEMENTS(size))
         return 0;
     return(size[dbr_type]);
 }
